+++ changelog.md
@@ -23,6 +23,7 @@ Codion Change Log
 - EntityDialogs bug fixed, now respects the location if specified.
 - EntityTablePanel, edit panel dialog location now follows mouse when displayed on double click.
 - EntityApplicationPanel.exitEvent() renamed exitObserver().
+- EntityDependenciesPanel no longer public, needs a better api.
 ### is.codion.framework.domain
 - DefaultKey.isNull(column) now throws exception in case the column is not part of the key.
 ### is.codion.framework.server